backend/requirements.txt
fastapi[standard]
pydantic
supabase
python-dotenv
uvicorn
httpx
geopy
pygeohash
